refactor: split ecstore rebalance unit tests
